refactor: move bulk download to new file
